Expose magnetometer and the LSM303 sensor
There are three major changes in this commit:

- Functionality for retrieving magnetometer data was added to the
  Accelerometer type, which prompted a renaming from Accelerometer to
  Sensor.
- This again prompted the renaming of the accelerometer module to
  motion.
- A separate function was added to the motion module that retrieves the
  Lsm303agr struct directly, instead of wrapping it in the Sensor
  struct.

Additionally, documentation was added to the methods and functions in
motion.
