+++ examples/accelerometer/src/main.rs
@@ -5,9 +5,9 @@ use defmt::{info, Debug2Format};
 use embassy_executor::Spawner;
 use embassy_time::Duration;
 use microbit_bsp::{
-    accelerometer::Accelerometer,
     display::{Brightness, Frame},
     embassy_nrf::{bind_interrupts, peripherals::TWISPI0, twim::InterruptHandler},
+    motion::Sensor,
     LedMatrix, Microbit,
 };
 use {defmt_rtt as _, panic_probe as _};
@@ -28,13 +28,13 @@ async fn main(_s: Spawner) {
     );
 
     let irqs = InterruptRequests {};
-    let mut acc = Accelerometer::new(board.twispi0, irqs, board.p23, board.p22).unwrap();
+    let mut sensor = Sensor::new(board.twispi0, irqs, board.p23, board.p22).unwrap();
 
-    let status = acc.accel_status().unwrap();
+    let status = sensor.accel_status().unwrap();
     info!("status: {:?}", Debug2Format(&status));
 
     loop {
-        let (x, y, z) = acc.accel_data().unwrap().xyz_mg();
+        let (x, y, z) = sensor.accel_data().unwrap().xyz_mg();
         #[allow(clippy::cast_precision_loss)]
         display_level(&mut display, Duration::from_millis(100), x as f32, y as f32, z as f32).await;
     }
