+++ src/lib.rs
@@ -5,12 +5,13 @@
 mod board;
 pub use board::*;
 
-pub mod accelerometer;
 pub mod display;
 pub mod mic;
+pub mod motion;
 pub mod speaker;
 
 // Re-exports
 
 pub use embassy_nrf;
 pub use lsm303agr;
+pub use nb;
